Clean up logic seperation between StorageRequest and S3Driver
